Fixed all warnings also with -ansi and -pedantic enabled in g++ 4.3.3. Warnings in generated files (Lua) and source included other projects (GLFT) where not fixed.

+++ src/StatsSystem.cpp
@@ -54,7 +54,7 @@ StatsSystem::~StatsSystem()
 	}
 }
 
-void StatsSystem::calculateTables( int maxLevel, double relativeIncreasePerLevel )
+void StatsSystem::calculateTables( size_t maxLevel, double relativeIncreasePerLevel )
 {
 	if ( maxStatsPointsForLevel != NULL ) {
 		assert( maxArmorPointsForLevel != NULL );
